% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apimSEM.R
\name{apimSEM}
\alias{apimSEM}
\title{A Function That Writes, Saves, and Exports Syntax for
Fitting Latent Actor-Partner Interdependence Models}
\usage{
apimSEM(dvn, lvxname, lvyname, model = "configural", compare = "none")
}
\arguments{
\item{dvn}{input object from dyadVarNames()}

\item{lvxname}{input character to (arbitrarily) name LV X in lavaan syntax}

\item{lvyname}{input character to (arbitrarily) name LV Y in lavaan syntax}

\item{model}{input character used to specify which level of invariance is
modeled. Defaults to "configural"}

\item{compare}{input character to specify which type of structural parameters
("actor" = actor effects, "partner" = partner effects,
"all_effects" = actor and partner effects, x_means, y_means, or all_means)
are constrained to equivalencey between partners. Default is "none".}
}
\value{
character object of lavaan script that can be passed immediately to
lavaan functions. Users will receive message if structural comparisons are specified
when the recommended level of invariance is not also specified.
}
\description{
This function takes the outputted object from dyadVarNames()
and automatically writes, returns, and exports (.txt) lavaan() syntax
for specifying Actor-Partner Interdependence Models (APIMs). Users can
also invoke configural, loading, and/or intercept invariant
measurement models, and particular types of structural comparisons.
}
\examples{
dvn = dyadVarNames(dat, xvar="X", yvar="Y", sep = ".",distinguish1 = "1", distinguish2 = "2")
apim.script.config = apimSEM(dvn, lvxname = "Conflict",
lvyname = "Satisfaction", model = "configural")
apim.script.load = apimSEM(dvn, lvxname = "Conflict",
lvyname = "Satisfaction", model = "loading")
apim.script.int = apimSEM(dvn, lvxname = "Conflict",
lvyname = "Satisfaction", model = "intercept")
apim.script.load.actor = apimSEM(dvn, lvxname = "Conflict",
lvyname = "Satisfaction", model = "loading", compare = "actor")
}
\seealso{
\code{\link{dyadVarNames}} which this function relies on
}
