% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bidyCFA.R
\name{bidyCFA}
\alias{bidyCFA}
\title{A Function That Writes, Saves, and Exports Syntax for
Fitting BiDY CFA Models}
\usage{
bidyCFA(dvn, lvname = "X", model = "configural")
}
\arguments{
\item{dvn}{input object from dyadVarNames()}

\item{lvname}{input character to (arbitrarily) name LV in lavaan syntax}

\item{model}{input character used to specify which level of invariance is
modeled ("configural", "loading", "intercept", "residual", or "indistinguishable"). Defaults to "configural".}

\item{lvnum}{optional input character to indicate which LV is modeled ("one" or "two").
Only necessary if dvn contains both X and Y information and user wants CFA for Y}
}
\value{
character object of lavaan script that can be passed immediately to
lavaan functions
}
\description{
This function takes the outputted object from dyadVarNames()
and automatically writes, returns, and exports (.txt) lavaan() syntax
for specifying dyadic configural, loading, and intercept invariant
BiDy measurement models for either a specified X or Y factor.
}
\examples{
dvn = dyadVarNames(dat, xvar="X", sep = ".",distinguish1 = "1", distinguish2 = "2")
con.bidy.config.script = bidyCFA(dvn, lvname = "Conflict", model = "configural")
con.bidy.loading.script = bidyCFA(dvn, lvname = "Conflict",  model = "loading")
con.bidy.intercept.script = bidyCFA(dvn, lvname = "Conflict",  model = "intercept")
}
\seealso{
\code{\link{dyadVarNames}} which this function relies on
}
