% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfa_funs.R
\name{cfa_funs}
\alias{cfa_funs}
\alias{loadings}
\alias{intercepts}
\alias{resids}
\title{Functions for dyadCFA() for scripting free or fixed families of parameters}
\usage{
loadings(dvn, fit)

intercepts(dvn, partner = "1", type = "free")

resids(dvn, partner = "1", type = "free")
}
\arguments{
\item{dvn}{input object from dyadVarNames()}

\item{partner}{input character to indicate parameters for first or second dyad member}

\item{type}{input character to indicate whether parameters "free" or "equated" in estimation}

\item{lvname}{input character to (arbitrarily) name LV in lavaan syntax}
}
\description{
Functions for dyadCFA() for scripting free or fixed families of parameters
}
\seealso{
\code{\link{dyadCFA}} which this function supplies
}
