% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commitmentM-data.R
\docType{data}
\name{commitmentM}
\alias{commitmentM}
\title{Ratings of relational satisfaction and commitment from 282 (M)ixed-sex couples}
\format{
A data frame with 282 rows and 20 variables:
\describe{
\item{sat.g1_f}{Satisfaction items for female partner}
\item{com1_f}{Commitment items items for female partner}
\item{sat.g1_m}{Satisfaction items for male partner}
\item{com1_m}{Commitment items items for male partner}
}
}
\usage{
data(commitmentM)
}
\description{
A data set containing 5 ratings of satisfaction and 5 ratings of commitment for each
member of a mixed-sex romantic dyad. Measured using the Investment Model Scale subscales
(Rubult, Martz, & Agnew, 1998). Data are from Sakaluk, Fisher, and Kilshaw's (in press) study
of dyadic invariance. Variable names have been re-coded to follow a stem-item-partner ("sip")
order, with a delimiter ("_") between the item number and partner distinguishing character.
}
\references{
Sakaluk, J. K., Fisher, A. N., & Kilshaw, R. E.(in press). Dyadic measurement invariance and its importance for replicability in romantic relationship research. \emph{Personal Relationships}.
}
\keyword{datasets}
