% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commitmentQ-data.R
\docType{data}
\name{commitmentQ}
\alias{commitmentQ}
\title{Ratings of relational satisfaction and commitment from 282 (Q)ueer couples}
\format{
A data frame with 118 rows and 20 variables:
\describe{
\item{sat.g.1_1}{Satisfaction items for female partner}
\item{com.1_1}{Commitment items items for female partner}
\item{sat.g.2_1}{Satisfaction items for male partner}
\item{com.2_1}{Commitment items items for male partner}
}
}
\usage{
data(commitmentQ)
}
\description{
A data set containing 5 ratings of satisfaction and 5 ratings of commitment for each
member of a same-sex romantic dyad. Measured using the Investment Model Scale subscales
(Rubult, Martz, & Agnew, 1998). Data are from analyses under review (Sakaluk, Fisher, and Kilshaw).
Variable names follow a stem-partner-item ("spi") order, with a delimiter (".") between
the stem and distinguishing partner character, and another delimiter ("_") between the
distinguishing partner character and item number.
}
\references{
Sakaluk, J. K., Fisher, A. N., & Kilshaw, R. E.(in press). Dyadic measurement invariance and its importance for replicability in romantic relationship research. \emph{Personal Relationships}.
}
\keyword{datasets}
