% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirs.R
\name{dirs}
\alias{dirs}
\title{A function for checking if subdirectories exist (and creating them if not)}
\usage{
dirs(type)
}
\arguments{
\item{type}{input character defining name of subdirectory corresponding to
type of outputted documents}
}
\description{
This function checks if subdirectories exist in the current working
directory, and creates them if not. Used in functions generating scripts,
figures, and tables
}
