% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRES-data copy.R
\docType{data}
\name{DRES}
\alias{DRES}
\title{Relationship quality and sexual satisfaction of 121 couples}
\format{
A data frame with 121 rows and 28 variables:
\describe{
\item{PRQC_1.1}{PRQC items for partner 1}
\item{PRQC_1.2}{PRQC items for partner 2}
\item{sexsat1.1}{sexual satisfaction items for partner 1}
\item{sexsat1.1}{sexual satisfaction items for partner 2}
}
}
\usage{
data(DRES)
}
\description{
A dataset containing 9 observed indicators of relationship quality (PRQC)
and 5 observed indicators of sexual satisfaction from 121 couples
in the DRES (Daily Relationship Experiences Study; Raposo & Muise, under review)
}
\references{
Raposo, S., & Muise, A. (Under review). Perceived Partner Sexual Responsiveness Buffers Anxiously Attached Individuals from Lower Relationship and Sexual Quality in Daily Life. Social Psychological and Personality Science.
}
\keyword{datasets}
