% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvn3.R
\name{dvn3}
\alias{dvn3}
\title{A Variable Name-Scraping and Indexing Function}
\usage{
dvn3(
  dat,
  x_order = "spi",
  time = NULL,
  x_stem,
  x_delim1 = NULL,
  x_delim2 = NULL,
  x_delim3 = NULL,
  x_item_num = "\\\\d+",
  distinguish_1 = "1",
  distinguish_2 = "2",
  y_order = NULL,
  y_stem = NULL,
  y_delim1 = NULL,
  y_delim2 = NULL,
  y_delim3 = NULL,
  y_item_num = "\\\\d+"
)
}
\arguments{
\item{dat}{input data frame of indicators of a particular LV}

\item{x_order}{input character for order of (S)tem, (P)artner number, (Item) number
and (T)ime (if applicable) when creating variable names. Defaults to "spit" (qualtrics-friendly)}

\item{time}{optional input numeric to specify number of longitudinal waves of names (defaults to null)}

\item{x_stem}{input character stem of indicator variables for LV X}

\item{x_delim1}{optional character to separate first name element (e.g., stem) from the second (e.g., item)}

\item{x_delim2}{optional character to separate second name element (e.g., item) from the third (e.g., partner)}

\item{x_delim3}{optional character to separate third name element (e.g., partner) from the fourth (e.g., time)
from final element of variable name}

\item{x_item_num}{defaults to scrape all items that match the stem with any digits that follow.
Will be updated to allow particular range of values, to make more sub-scale friendly.}

\item{y_order}{optional character for order of (S)tem, (P)artner number, and (Item) number
when creating variable names. Defaults to "spi" (qualtrics-friendly). This and other Y-arguments
only necessary if there is a latent Y variable to model}

\item{y_stem}{optional input character stem of indicator variables for LV X}

\item{y_delim1}{optional character to separate stem from partner number (spi) or item number (sip)}

\item{y_delim2}{optional character to separate stem/partner number (spi) or stem/item number (sip) from
from final element of variable name}

\item{y_item_num}{defaults to scrape all items that match the stem with any digits that follow.
Will be updated to allow particular range of values, to make more sub-scale friendly.}

\item{distinguish1}{input character used as the identifier for the first partner}

\item{distinguish2}{input character used as the identifier for the first partner}
}
\value{
list containing variable names for p1, p2, # of items per LV,
characters distinguishing partners, and total number of indicators
}
\description{
This function scrapes the names of indicator variables in a
wide-format data set used for dyadic analyses of two latent variables (LV; X and Y),
and indexes which indicators correspond to which partner, for which LV.
It is used primarily to guide the syntax-writing of the other dySEM functions.
}
\examples{
dvn <- dvn2(dat = dat.pkg, x_order = "sip", x_stem = "X", x_delim2=".", x_item_num="\\\\d+", distinguish_1="1", distinguish_2="2",
y_order="sip", y_stem="Y", y_delim2=".", y_item_num="\\\\d+")
}
