% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyadVarNames.R
\name{dyadVarNames}
\alias{dyadVarNames}
\title{A Variable Name-Scraping and Indexing Function}
\usage{
dyadVarNames(dat, xvar = NULL, yvar = NULL, sep, distinguish1, distinguish2)
}
\arguments{
\item{dat}{input data frame of indicators of a particular LV}

\item{xvar}{input character stem of indicator variables for LV X}

\item{yvar}{input character stem of indicator variables for LV Y}

\item{sep}{input character use to separate variable names from the partner identifier}

\item{distinguish1}{input character used as the identifier for the first partner}

\item{distinguish2}{input character used as the identifier for the first partner}
}
\value{
list containing variable names for p1, p2, # of items per LV,
characters distinguishing partners, and total number of indicators
}
\description{
This function scrapes the names of indicator variables in a
wide-format data set used for dyadic analyses of two latent variables (LV; X and Y),
and indexes which indicators correspond to which partner, for which LV.
It is used primarily to guide the syntax-writing of the other dySEM functions.
}
\examples{
dvn = dyadVarNames(dat, xvar="X", yvar="Y", sep = ".",distinguish1 = "1", distinguish2 = "2")
}
