% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dydMACS.R
\name{dydMACS}
\alias{dydMACS}
\title{A Function That Computes a Version of the d_MACS Effect Size Measure
of Noninvariance (see Nye & Drasgow, 2011; Nye et al., 2018).}
\usage{
dydMACS(fit.cfa)
}
\arguments{
\item{fit.cfa}{output from a dyadCFA() configural invariance model fitted with
lavaan()}
}
\value{
numeric vector of dMACS values for each indicator of a given LV
}
\description{
This function takes a model written with dyadCFA() and fitted with lavaan()
and returns d_MACS values for each indicator variable. Function is a renamed
version of my  dMACS_long function (https://github.com/jsakaluk/dMACS_long),
which was adapted from Johannes Karl's (https://johanneskarl.netlify.com) between-groups
dMACS function (https://github.com/Jo-Karl/dMACS) with his assistance.
}
\examples{
dvn = dyadVarNames(dat, xvar="X", sep = ".",distinguish1 = "1", distinguish2 = "2")
con.config.script = dyadCFA(dvn, lvname = "Conflict",  model = "configural")
con.fit.config <- lavaan::cfa(con.config.script, data = dat, std.lv = FALSE,
auto.fix.first= FALSE, meanstructure = TRUE)
dydMACS(con.fit.config)
}
\seealso{
\code{\link{dyadCFA}} which this function relies on
}
