% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyInvarES.R
\name{dyInvarEs}
\alias{dyInvarEs}
\title{A Function That Computes Effect Sizes of Dyadic Noninvariance}
\usage{
dyInvarEs(fit.cfa)
}
\arguments{
\item{fit.cfa}{output from a dyadCFA() configural invariance model fitted with
lavaan()}
}
\value{
numeric vector of dMACS values for each indicator of a given LV
}
\description{
This function takes a model written with dyadCFA() and fitted with lavaan()
and returns d_MACS values for each indicator variable. Function is a renamed
version of my  dMACS_long function (https://github.com/jsakaluk/dMACS_long),
which was adapted from Johannes Karl's (https://johanneskarl.netlify.com) between-groups
dMACS function (https://github.com/Jo-Karl/dMACS) with his assistance.
}
\examples{
dvn <- dvn2(DRES, x_order = "sip", x_stem = "sexsat", x_delim2=".", x_item_num="\\\\d+", distinguish_1="1", distinguish_2="2")
config.script = dyadCFA(dvn, lvname = "SexSat",  model = "configural")
fit <- lavaan::cfa(config.script, data = DRES, std.lv = FALSE,
auto.fix.first= FALSE, meanstructure = TRUE)
dydMACS(con.fit.config)
}
\seealso{
\code{\link{dyadCFA}} which this function relies on
}
