% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyMIOutput.R
\name{dyMIOutput}
\alias{dyMIOutput}
\title{A Function That Exports a Table of Dyadic Invariance Model Comparison Indices}
\usage{
dyMIOutput(config, load, int, robust = F)
}
\arguments{
\item{config}{fitted dyadic configural invariance model lavaan object}

\item{load}{fitted dyadic loading invariance model lavaan object}

\item{int}{fitted dyadic intercept invariance model lavaan object}

\item{robust}{logical for whether a robust estimator has been used (default is F)}
}
\value{
Prints word files for the table outputed to the users working directory
}
\description{
This function takes the model from fitted dyCFA() scripts and
exports a table of expected output.
}
\examples{
dvn <- dyadVarNames(dat, xvar="X",  sep = ".",
distinguish1 = "1", distinguish2 = "2")

con.config.script = dyadCFA(dvn, lvname = "Conflict",  model = "configural")
con.loading.script = dyadCFA(dvn, lvname = "Conflict",  model = "loading")
con.intercept.script = dyadCFA(dvn, lvname = "Conflict",  model = "intercept")

con.fit.config <- lavaan::cfa(con.config.script, data = dat, std.lv = FALSE, auto.fix.first= FALSE, meanstructure = TRUE)
con.fit.loading <- lavaan::cfa(con.loading.script, data = dat, std.lv = FALSE, auto.fix.first= FALSE, meanstructure = TRUE)
con.fit.intercept <- lavaan::cfa(con.intercept.script, data = dat, std.lv = FALSE, auto.fix.first= FALSE, meanstructure = TRUE)

dyMIOutput(con.fit.config, con.fit.loading, con.fit.intercept, robust = FALSE)
}
\seealso{
\code{\link{dyadCFA}} which this function relies on
}
