% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyRand.R
\name{dyRand}
\alias{dyRand}
\title{A Function that randomizes the position (1st or 2nd) of dyad members' observations within a row of a data frame}
\usage{
dyRand(dvn, dat)
}
\arguments{
\item{dvn}{input object from dyadVarNames()}

\item{dat}{input data frame}
}
\value{
A randomly permuted data frame
}
\description{
This function takes the variables in a given dvn, corresponding to a data frame
and creates a randomly permuted version of the data frame wherein each dyad has
the column positions for each member within a row randomly determined to come first or last.
This is a helper function for the randomization permutation testing function of
dyadic invariance for ostensibly indistinguishable dyads.
}
\examples{
dat = dplyr::select(DRES, sexsat1.1:sexsat5.2)
dvn = dyadVarNames(dat, xvar = "sexsat", sep = ".", distinguish1 = "1", distinguish2 = "2")
perm = dyRand(dvn, dat)
}
\seealso{
\code{\link{dyadVarNames}} which this function relies on
}
