% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyRandPermFit.R
\name{dyRandPermFit}
\alias{dyRandPermFit}
\title{A Function that fits dyadic CFA models to randomly permuted data frames}
\usage{
dyRandPermFit(dvn, dat, n, AFIs = NULL, model = "configural")
}
\arguments{
\item{dvn}{input object from dyadVarNames()}

\item{dat}{input data frame}

\item{n}{input numeric argument for number of randomly permuted samples to use}

\item{AFIs}{Optional character vector indicating which alternative fit indices (or chi-squared itself)}

\item{model}{Character argument indicating type of invariance model to evaluate
are to be used to be extracted from the model fit in each permuted sample}
}
\value{
A randomly permuted data frame
}
\description{
This function fits  dyadic CFA models to randomly permuted data frames
generated via dyRand(), and extracts specified fit indexes from these null cases to
form an empirical sampling distribution to be used for testing dyadic invariance
for indistinguishable dyads
}
\examples{
dat = dplyr::select(DRES, sexsat1.1:sexsat5.2)
dvn = dyadVarNames(dat, xvar = "sexsat", sep = ".", distinguish1 = "1", distinguish2 = "2")
permout = dyRandPermFit(dvn, dat, n = 10)
}
\seealso{
\code{\link{dyadVarNames}} \code{\link{dyRand}} which this function relies on
}
