% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyRandPermTest.R
\name{dyRandPermTest}
\alias{dyRandPermTest}
\title{A performs random permuted tests of  invariance for ostensibly indistinguishable  dyads}
\usage{
dyRandPermTest(dvn, dat, n, AFIs = NULL, model = "configural")
}
\arguments{
\item{dvn}{input object from dyadVarNames()}

\item{dat}{input data frame}

\item{n}{input numeric argument for number of randomly permuted samples to use}

\item{AFIs}{Optional character vector indicating which alternative fit indices (or chi-squared itself)}

\item{model}{Character argument indicating type of invariance model to evaluate
are to be used to be extracted from the model fit in each permuted sample}
}
\value{
A randomly permuted data frame
}
\description{
This function fits  dyadic CFA models to randomly permuted data frames
generated via dyRand() and dyRandPermFit(), and presents a numeric and visual
comparison of the fit of the observed invariance model against
the fit of the empirical null sampling distribtuion
}
\examples{
dat = dplyr::select(DRES, sexsat1.1:sexsat5.2)
dvn = dyadVarNames(dat, xvar = "sexsat", sep = ".", distinguish1 = "1", distinguish2 = "2")
dyRandPermTest(dvn, dat, n = 10, model = "configural")
}
\seealso{
\code{\link{dyadVarNames}} \code{\link{dyRand}} \code{\link{dyRandPermFit}} which this function relies on
}
