% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyReliability.R
\name{dyReliability}
\alias{dyReliability}
\title{A Function Calculates Omega Total Coefficients from a Dyadic CFA}
\usage{
dyReliability(dvn, fit)
}
\arguments{
\item{dvn}{input object from dyadVarNames()}

\item{fit}{input object from fitted lavaan model based on the default (i.e., "configural") dyadCFA() function}
}
\value{
Calculated omega total coefficients for dyad Member 1 and Member 2
}
\description{
This function takes the  model from fitted dyCFA() scripts and returns omega total coefficients
for each dyad member following Formula 2 in McNeish (2018).
}
\examples{
dvn = dyadVarNames(dat, xvar="X", sep = ".",distinguish1 = "1", distinguish2 = "2")
con.config.script = dyadCFA(dvn, lvname = "Conflict", model = "configural")
con.config.mod <- cfa(con.config.script, data = dat, std.lv = F, auto.fix.first= F, meanstructure = T)
dyReliability(con.config.mod)
}
