% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dydmacs.R
\name{get_dydmacs}
\alias{get_dydmacs}
\title{Calculates dmacs difference in expected indicator scores for between dyad members}
\usage{
get_dydmacs(dat, fit, nodewidth = 0.01, lowerLV = -5, upperLV = 5)
}
\arguments{
\item{dat}{data frame of indicators}

\item{fit}{outputted dyadic cfa lavaan object}

\item{nodewidth}{space between nodes during quadrature approximation (default = .01)}

\item{lowerLV}{lowest latent variable value evaluated (default = -5)}

\item{upperLV}{greatest latent variable value evaluated (default = 5)}
}
\value{

}
\description{
Calculates dmacs difference in expected indicator scores for between dyad members
}
\examples{
dat <-  DRES
dvn <- dvn2(dat, x_order = "sip", x_stem = "sexsat", x_delim2=".", x_item_num="\\\\d+", distinguish_1="1", distinguish_2="2")
config.script<- dyadCFA(dvn, lvname = "SexSat",  model = "configural")
fit <- lavaan::cfa(config.script, data = dat, std.lv = FALSE,
auto.fix.first= FALSE, meanstructure = TRUE)
sexsat.dmacs <- get_dydmacs(dat, fit)
}
