% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDydmacs.R
\name{getDydmacs}
\alias{getDydmacs}
\title{Calculates dmacs difference in expected indicator scores for between dyad members}
\usage{
getDydmacs(dat, dvn, fit, nodewidth = 0.01, lowerLV = -5, upperLV = 5)
}
\arguments{
\item{dat}{data frame of indicators}

\item{dvn}{dvn for the dyadic cfa measurement model}

\item{fit}{outputted dyadic cfa lavaan object; should be from a partial-invariance model}

\item{nodewidth}{space between nodes during quadrature approximation (default = .01)}

\item{lowerLV}{lowest latent variable value evaluated (default = -5)}

\item{upperLV}{greatest latent variable value evaluated (default = 5)}
}
\value{

}
\description{
Calculates dmacs difference in expected indicator scores for between dyad members
}
\examples{
dvn <- scrapeVarCross(dat = DRES, x_order = "sip", x_stem = "PRQC", x_delim1 = "_", x_delim2=".", x_item_num="\\\\d+", distinguish_1="1", distinguish_2="2")
qual.config.script <-  scriptCFA(dvn, lvname = "Qual", model = "configural")
qual.config.mod <- cfa(qual.config.script, data = DRES, std.lv = F, auto.fix.first= F, meanstructure = T)
qual.dmacs <- getDydmacs(DRES, dvn, qual.config.mod)

}
\seealso{
Other supplemental model calculators: 
\code{\link{getDyReliability}()},
\code{\link{getIndistFit}()}
}
\concept{supplemental model calculators}
