% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIndistFit.R
\name{getIndistFit}
\alias{getIndistFit}
\title{A Function that Computes Corrected Fit Indexes According to the ISAT and INULL Models of Olsen & Kenny (2006)}
\usage{
getIndistFit(indmodel, isatmod, inullmod)
}
\arguments{
\item{indmodel}{input lavaan model fitted using dyadCFA(model = "indistinguishable")}

\item{isatmod}{input lavaan model fitted using ISAT()}

\item{inullmod}{input lavaan model fitted using INULL()}
}
\value{
A data frame of the original and corrected chi sq, df, p, rmsea, and tli
}
\description{
This function takes the outputted model fit using scriptCFA() with model = "indist",
as well as scriptISAT(), and scriptINULL() and computes corrected model fit indexes according
to the approach outlined by Olsen & Kenny (2006)
}
\examples{
dvn <- scrapeVarCross(dat = DRES, x_order = "sip", x_stem = "PRQC", x_delim1 = "_", x_delim2=".", x_item_num="\\\\d+", distinguish_1="1", distinguish_2="2")
qual.indist.script <-  scriptCFA(dvn, lvname = "Qual", model = "indist")
qual.indist.mod <- cfa(qual.indist.script, data = DRES, std.lv = F, auto.fix.first= F, meanstructure = T)
qual.isat.script <- scriptISAT(dvn, lvxname = "Qual")
qual.isat.mod <- cfa(qual.isat.script, data = DRES, std.lv = F, auto.fix.first= F, meanstructure = T)
qual.inull.script <- scriptINULL(dvn, lvxname = "Qual")
qual.inull.mod <- cfa(qual.isat.script, data = DRES, std.lv = F, auto.fix.first= F, meanstructure = T)
corr.fit <- getIndistFit(qual.indist.mod, qual.isat.mod, qual.inull.mod)

}
\seealso{
Other supplemental model calculators: 
\code{\link{getDyReliability}()},
\code{\link{getDydmacs}()}
}
\concept{supplemental model calculators}
