% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUniConstructMods.R
\name{getUniConstructMods}
\alias{getUniConstructMods}
\title{A Function That Fits and Examines Competing Dyadic Uni-construct Models}
\usage{
getUniConstructMods(
  dvn,
  dat,
  indexes = c("npar", "df", "chisq", "cfi", "rmsea", "tli", "aic", "bic"),
  modelCompare = TRUE,
  genTEFI = TRUE,
  ...
)
}
\arguments{
\item{dvn}{Input dvn list from \code{scrapeVarCross()}.}

\item{dat}{Input data frame containing the dataset for model estimation.}

\item{indexes}{Input character vector specifying which index(es) to return.
Default is \code{c("npar", "df", "chisq", "cfi", "rmsea", "tli", "aic", "bic")}.
\strong{Note}: Must be valid options from \code{lavaan::fitMeasures()}.}

\item{modelCompare}{A logical input indicating whether to compare fit statistics
(i.e., chi-squared, AIC, and BIC) of the fitted models. The default is \code{TRUE}.
\strong{Note}: A scaled chi-squared difference test will automatically be applied
(via \code{lavaan::lavTestLRT()}) if a robust estimator variant (e.g., \code{estimator = "mlr"})
is used.}

\item{genTEFI}{A logical input indicating whether to compute the
Generalized Total Entropy Fit Index (genTEFI) (see Golino et al., 2024).
Default is \code{TRUE}.}

\item{...}{Additional arguments to be passed to \code{lavaan::cfa()}, allowing
users to customize model estimation settings. By default, the models will
be fit with robust maximum-likelihood (MLR) estimation---\code{estimator = "mlr"}
---and missing data will handled using full-information maximum likelihood
(FIML)---\code{missing = "fiml"}.}
}
\value{
A \code{list} containing up to three components:
\itemize{
\item \verb{Global Fit}: A \code{tibble} with the desired index(es) for each fitted model (requested via the \code{indexes} argument).
\item \verb{Model Comparisons}: A \code{tibble} summarizing the differences in fit statistics between the fitted models. (if \code{modelCompare = TRUE}).
\item \code{genTEFI}: A \code{tibble} of the genTEFI (if \code{genTEFI = TRUE}).
}
}
\description{
This function takes the outputted object from \code{scrapeVarCross()}
along with the corresponding dataset and automatically tests competing
uni-construct dyadic models for the latent variable under consideration.
It inspects four possible model variants:
\itemize{
\item Unidimensional (via \code{scriptUni})
\item Correlated Factors (via \code{scriptCor})
\item Hierarchical (via \code{scriptHier})
\item Bifactor (via \code{scriptBifac})
}
}
\examples{

dvn <- scrapeVarCross(
  commitmentQ,
  x_order = "spi",
  x_stem = "sat.g",
  x_delim1 = ".",
  x_delim2="_",
  distinguish_1="1",
  distinguish_2="2"
  )
  
  getUniConstructMods(
  dvn,
  commitmentQ
  )
  
  getUniConstructMods(
  dvn,
  commitmentQ,
  indexes = c("npar", "df", "bic"),
  modelCompare = FALSE,
  genTEFI = TRUE
  )
  
  getUniConstructMods(
  dvn,
  commitmentQ,
  indexes = c("npar", "df", "bic"),
  modelCompare = FALSE,
  genTEFI = TRUE,
  estimator = "ml"
  )

}
