% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indistFit.R
\name{indistFit}
\alias{indistFit}
\title{A Function that Computes Corrected Fit Indexes According to the ISAT and INULL Models of Olsen & Kenny (2006)}
\usage{
indistFit(indmodel, isatmod, inullmod)
}
\arguments{
\item{indmodel}{input lavaan model fitted using dyadCFA(model = "indistinguishable")}

\item{isatmod}{input lavaan model fitted using ISAT()}

\item{inullmod}{input lavaan model fitted using INULL()}
}
\value{
A data frame of the original and corrected chi sq, df, p, rmsea, and tli
}
\description{
This function takes the outputted object from dyadVarNames()
and automatically writes, returns, and exports (.txt) lavaan() syntax
for specifying dyadic configural, loading, and intercept invariant
measurement models for either a specified X or Y factor.
}
\examples{
dvn = dyadVarNames(dat, xvar="X", sep = ".",distinguish1 = "1", distinguish2 = "2")
indist.script <- dyadCFA(dvn, lvname = "Conflict", model = "indistinguishable")
indist.fit <- cfa(indist.script, data = dat, std.lv = F, auto.fix.first= F, meanstructure = T)
isat.mod <- ISAT(dvn, lvname = "Conflict")
isat.fit <- cfa(isat.mod, data = dat, std.lv = F, auto.fix.first= F, meanstructure = T)
inull.mod <- INULL(dvn, lvname = "Conflict")
inull.fit <- cfa(inull.mod, data = dat, std.lv = F, auto.fix.first= F, meanstructure = T)
indistFit(indmodel = indist.fit, isatmod = isat.fit, inullmod = inull.fit)
}
