% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISAT.R
\name{ISAT}
\alias{ISAT}
\title{A Function That Writes, Saves, and Exports Syntax for
Fitting the I-SAT model for indistinguishable dyads}
\usage{
ISAT(dvn, lvname)
}
\arguments{
\item{dvn}{input object from dyadVarNames()}

\item{lvname}{input character to (arbitrarily) name LV in lavaan syntax}
}
\value{
character object of lavaan script that can be passed immediately to
lavaan functions
}
\description{
This function takes the outputted object from dyadVarNames()
and automatically writes, returns, and exports (.txt) lavaan() syntax
for specifying dyadic configural, loading, and intercept invariant
measurement models for either a specified X or Y factor.
}
\examples{
dvn = dyadVarNames(dat, xvar="X", sep = ".",distinguish1 = "1", distinguish2 = "2")
con.config.script = dyadCFA(dvn, lvname = "Conflict", model = "configural")
con.loading.script = dyadCFA(dvn, lvname = "Conflict",  model = "loading")
con.intercept.script = dyadCFA(dvn, lvname = "Conflict",  model = "intercept")
}
\seealso{
\code{\link{dyadVarNames}} which this function relies on
}
