% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaanscript_funs.R
\name{lavaanscript_funs}
\alias{lavaanscript_funs}
\alias{x1_loads}
\title{Functions for scripting portions of lavaan() code.}
\usage{
x1_loads(dvn, lvxname = lvname)
}
\arguments{
\item{dat}{input data frame of indicators of a particular LV}

\item{xvar}{input character stem of indicator variables for LV X}

\item{yvar}{input character stem of indicator variables for LV Y}

\item{index1}{combined character of sep and identifier for the first partner}

\item{index2}{combined character of sep and identifier for the first partner}
}
\description{
Functions for scripting portions of lavaan() code.
}
\seealso{
\code{\link{dyadVarNames}} which this function supplies
}
