% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurement_output.R
\name{measurement_funs}
\alias{measurement_funs}
\alias{loadings}
\alias{xintercepts}
\alias{yintercepts}
\title{Functions for dyOutput() for extracting measurement model output}
\usage{
loadings(dvn, fit)

xintercepts(dvn, fit)

yintercepts(dvn, fit)
}
\arguments{
\item{dvn}{input object from dyadVarNames()}

\item{fit}{input object from fitted lavaan model}
}
\description{
Functions for dyOutput() for extracting measurement model output
}
\seealso{
\code{\link{dyOutput}} which this function supplies
}
