% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outHelpers.R
\name{outHelpers}
\alias{outHelpers}
\alias{makeTable}
\alias{makeFigure}
\title{Helpers for creating output files}
\usage{
makeTable(dvn, fit, model, tabletype)

makeFigure(fit, type)
}
\arguments{
\item{dvn}{input dvn list from scrapeVarCross}

\item{fit}{name of fitted lavaan model}

\item{model}{type of fitted dyadic model (i.e., "apim", "bidyc", "bidys","cfa", "cfm", "mim")}

\item{tabletype}{kind of parameter estimates requested (i.e. from "measurement" or "structural" model)}

\item{type}{input character for sempaths to indicate whether parameters "free" or "equated" in estimation}
}
\description{
Helpers for creating output files
}
\seealso{
Other helpers: 
\code{\link{esHelpers}},
\code{\link{outParamHelpers}},
\code{\link{scrapeHelpers}},
\code{\link{scriptHelpers}},
\code{\link{semHelpers}}
}
\concept{helpers}
