% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputInvarItem.R
\name{outputInvarItem}
\alias{outputInvarItem}
\title{Output Item-Level Invariance Summary Table}
\usage{
outputInvarItem(
  dvn,
  fit,
  partialScript,
  dat,
  nodewidth = 0.01,
  lowerLV = -5,
  upperLV = 5,
  gtTab = FALSE,
  writeTo = NULL,
  fileName = NULL
)
}
\arguments{
\item{dvn}{input dvn list from scrapeVarCross}

\item{fit}{outputted dyadic cfa lavaan object from the partial invariance model}

\item{partialScript}{character string containing the lavaan syntax for the partial invariance model}

\item{dat}{data frame of indicators used to fit the model (needed for dMACS calculation)}

\item{nodewidth}{space between nodes during quadrature approximation for dMACS (default = .01)}

\item{lowerLV}{lowest latent variable value evaluated for dMACS (default = -5)}

\item{upperLV}{greatest latent variable value evaluated for dMACS (default = 5)}

\item{gtTab}{A logical input indicating whether to generate the output in \code{gt::gt()} table object format (\code{TRUE}).
By default (\code{FALSE}), the output is generated in \code{tibble::tibble()} format.
Users can also apply the \code{writeTo} argument if they wish to export the \code{gt:gt()} table object.}

\item{writeTo}{A character string specifying a directory path to where the \code{gt::gt()} table object should be saved.
If set to ".", the file will be written to the current working directory.
The default is \code{NULL}, and examples use a temporary directory created by \code{tempdir()}.
\code{writeTo} is only relevant if \code{gtTab = TRUE}.}

\item{fileName}{A character string specifying a desired base name for the output \code{gt::gt()} file.}
}
\value{
A \code{tibble::tibble()} if \code{gtTab = FALSE} (default), or \code{gt::gt()} object if \code{gtTab = TRUE}, with
columns for Item, LoadingNoninvariance, InterceptNoninvariance, and dMACS. For multi-LV structures,
an additional LV column is included; in gt tables, this is used as the row group label with Item names
as row labels within each group.
}
\description{
\code{outputInvarItem()} generates a summary table showing which items exhibit
noninvariance in loadings and/or intercepts based on a partial invariance model script,
along with dMACS effect sizes for each item.
}
\details{
\itemize{
\item The function parses the lavaan syntax to determine which loadings and intercepts are constrained
to equality across partners (invariant) versus freed (noninvariant).
\item LoadingNoninvariance = "Yes" indicates the loading was freed in the partial invariance model.
\item InterceptNoninvariance = "Yes" indicates the intercept was freed in the partial invariance model.
\item dMACS values are calculated using the \code{getDydmacs()} function.
\item If \code{gtTab = TRUE} and \code{writeTo} is specified, then output will simultaneously be saved as a \code{.rtf} file to the user's specified directory.
\item If output file is successfully saved, a confirmation message will be printed to the console.
\item If a file with the same name already exists in the user's chosen directory, it will be overwritten.
}
}
\examples{
\dontrun{
dvn <- scrapeVarCross(dat = commitmentQ, x_order = "spi", x_stem = "sat.g", x_delim1 = ".",
                      x_delim2 = "_", distinguish_1 = "1", distinguish_2 = "2")

# Create and fit a partial invariance model
partial.script <- scriptCor(dvn, lvname = "Sat",
                            constr_dy_meas = c("loadings", "intercepts"),
                            constr_dy_struct = "none")

partial.fit <- lavaan::cfa(partial.script, data = commitmentQ, std.lv = FALSE,
                           auto.fix.first = FALSE, meanstructure = TRUE)

outputInvarItem(dvn, partial.fit, partial.script, dat = commitmentQ,
                gtTab = TRUE, writeTo = tempdir(), fileName = "item_invar")
}

}
\concept{output functions}
