% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputModel.R
\name{outputModel}
\alias{outputModel}
\title{A Function That Exports Tables and/or SEM Digrams based on dySEM models}
\usage{
outputModel(
  dvn,
  model = NULL,
  fit,
  table = TRUE,
  tabletype = NULL,
  figure = TRUE,
  figtype = NULL
)
}
\arguments{
\item{dvn}{input object from dyadVarNames()}

\item{model}{character input specying type of model to output (e.g., "cfa", "apim", "cfm")}

\item{fit}{input object from fitted lavaan model}

\item{table}{logical input of whether table output is desired. Default is TRUE}

\item{tabletype}{character input of what type of table(s) is(are) desired.
options are "measurement" (i.e,, slopes, intercepts, and (optionally) dMACS),
"structural" (i.e., latent slopes, such as actor/partner effects, k parameters), or "both"
(i.e., both measurement and structural tables)}

\item{figure}{logical input of whether figure output is desired. Default is TRUE}

\item{figtype}{character input of what type of figure is desired.}
}
\value{
Ignore console (prints unneccessary semPlot::semPaths details). More importantly,
prints word files for the table(s) and/or figure, outputed to the users working directory
}
\description{
This function takes the  model from fitted dySEM() scripts and
exports table(s) and/or a path diagram figure of expected output.
}
\examples{
dvn = dyadVarNames(dat, xvar="X", yvar="Y", sep = ".",
distinguish1 = "1", distinguish2 = "2")
apim.script.load = apimSEM(dvn, lvxname = "Conflict",
lvyname = "Satisfaction", model = "loading")
apim.fit.load <- lavaan::cfa(apim.script.load, data = dat, std.lv = FALSE,
auto.fix.first= FALSE, meanstructure = TRUE)
dyOutput(dvn, model = "cfa", fit, figure = FALSE, dydMACS.x = sexsat.dmacs)
}
\seealso{
\code{\link{dyadVarNames}} \code{\link{apimSEM}} \code{\link{dydMACS}} which this function relies on
}
