% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeHelpers.R
\name{scrapeHelpers}
\alias{scrapeHelpers}
\alias{sipExtractor}
\alias{spiExtractor}
\alias{spitExtractor}
\alias{siptExtractor}
\title{Helper-functions for scraping variable names and data set properties}
\usage{
sipExtractor(dat, stem, delim1, item_num, delim2, distinguish)

spiExtractor(dat, stem, delim1, item_num, delim2, distinguish)

spitExtractor(dat, stem, delim1, item_num, delim2, distinguish, delim3, wave)

siptExtractor(dat, stem, delim1, item_num, delim2, distinguish, delim3, wave)
}
\arguments{
\item{dat}{input data frame of indicators of a particular LV}

\item{stem}{input character stem of indicator variables for LV}

\item{delim1}{optional input character of first delimiting character}

\item{item_num}{optional input character to indicate range of variable numbers (defaults to scrape any digits)}

\item{delim2}{optional input character of second delimiting character}

\item{distinguish}{input character used as the identifier a particular partner}

\item{delim3}{optional input character of second delimiting character}

\item{wave}{input character for the particular wave of a longitudinal design}
}
\description{
Helper-functions for scraping variable names and data set properties
}
\seealso{
Other helpers: 
\code{\link{esHelpers}},
\code{\link{outHelpers}},
\code{\link{outParamHelpers}},
\code{\link{scriptHelpers}},
\code{\link{semHelpers}}
}
\concept{helpers}
