% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scriptAPIM.R
\name{scriptAPIM}
\alias{scriptAPIM}
\title{A Function That Writes, Saves, and Exports Syntax for
Fitting Latent Actor-Partner Interdependence Models}
\usage{
scriptAPIM(
  dvn,
  scaleset = "FF",
  lvxname,
  lvyname,
  model = "indist",
  equate = "none",
  k = FALSE,
  writescript = FALSE
)
}
\arguments{
\item{dvn}{input dvn list from scrapeVarCross}

\item{scaleset}{input character to specify how to set the scale of the latent variable(s). Default is
"FF" (fixed-factor; see Details for rationale), but user can specify "MV" (Marker Variable)}

\item{lvxname}{input character to (arbitrarily) name LV X in lavaan syntax}

\item{lvyname}{input character to (arbitrarily) name LV Y in lavaan syntax}

\item{model}{input character used to specify which level of invariance is
modeled. Defaults to "configural"}

\item{equate}{input character to specify which type of structural parameters
("actor" = actor effects, "partner" = partner effects,
"all_effects" = actor and partner effects, x_means, y_means, or all_means)
are constrained to equivalencey between partners. Default is "none". Actor or partner effect constraints
require at least a loading-invariant model to be specified, otherwise a warning is returned; mean constraints
require an intercept-invariant model to be specified, otherwise a warning is returned.}

\item{k}{input logical for whether Kenny & Ledermann's (2010) k parameter should be
calculated to characterize the dyadic pattern in the APIM. Defaults FALSE, and requires at least
a loading-invariant model to be specified, otherwise a warning is returned.}

\item{writescript}{input logical (default FALSE) for whether lavaan script should
be concatenated and written to current working directory (in subdirectory "scripts")}
}
\value{
character object of lavaan script that can be passed immediately to
lavaan functions. Users will receive message if structural comparisons are specified
when the recommended level of invariance is not also specified. If user supplies dvn
with containing X or Y variables, they are alerted to respecify the dvn object.
}
\description{
This function takes the outputted object from dyadVarNames()
and automatically writes, returns, and exports (.txt) lavaan() syntax
for specifying Actor-Partner Interdependence Models (APIMs). Users can
also invoke configural, loading, and/or intercept invariant
measurement models, and particular types of structural comparisons.
}
\examples{
dvn <- scrapeVarCross(dat = commitmentQ, x_order = "spi", x_stem = "sat.g", x_delim1 = ".", x_delim2="_", distinguish_1="1", distinguish_2="2",
y_order="spi", y_stem="com", y_delim1 = ".", y_delim2="_")
apim.script.config <-  scriptAPIM(dvn, lvxname = "Sat", lvyname = "Com", model = "configural", k = TRUE)
apim.script.load <- scriptAPIM(dvn, lvxname = "Sat", lvyname = "Com", model = "loading")
apim.script.int  <- scriptAPIM(dvn, lvxname = "Sat", lvyname = "Com", model = "intercept")
apim.script.ind <- scriptAPIM(dvn, lvxname = "Sat", lvyname = "Com", model = "indist")
}
\seealso{
\code{\link{dyadVarNames}} which this function relies on

Other script-writing functions: 
\code{\link{scriptBiDy}()},
\code{\link{scriptCFA}()},
\code{\link{scriptCFM}()},
\code{\link{scriptINULL}()},
\code{\link{scriptISAT}()},
\code{\link{scriptMIM}()}
}
\concept{script-writing functions}
