% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scriptBiDy.R
\name{scriptBiDy}
\alias{scriptBiDy}
\title{A Function That Writes, Saves, and Exports Syntax for
Fitting BiDY  Models}
\usage{
scriptBiDy(dvn, lvname = "X", type = "C", model = "configural")
}
\arguments{
\item{dvn}{input dvn list from scrapeVarCross}

\item{lvname}{input character to (arbitrarily) name LV in lavaan syntax}

\item{type}{input character used to specify what kind of BiDy model to specify ("C" for BiDy-C, or "S" for BiDy-S)}

\item{model}{input character used to specify which level of invariance is
modeled ("configural", "loading", "intercept", "residual", or "indistinguishable"). Defaults to "configural".}
}
\value{
character object of lavaan script that can be passed immediately to
lavaan functions
}
\description{
This function takes the outputted object from scrapeVarCross()
and automatically writes, returns, and exports (.txt) lavaan() syntax
for specifying dyadic configural, loading, and intercept invariant
BiDy CFA (BiDy-C) or SEM (BiDy-S) Model. Currently only uses fixed-factor scale-setting
}
\examples{
dvn <- scrapeVarCross(DRES, x_order = "sip", x_stem = "sexsat",  x_delim2=".", distinguish_1="1", distinguish_2="2")
sexsat.bidy.config.script <- scriptBiDy(dvn, lvname = "SexSat", model = "configural", type = "C")
sexsat.bidy.loadsource.script <- scriptBiDy(dvn, lvname = "SexSat", model = "loading_source", type = "C")
sexsat.bidy.loadreleq.script <- scriptBiDy(dvn, lvname = "SexSat", model = "loading_releq", type = "C")
}
\seealso{
\code{\link{dyadVarNames}} which this function relies on

Other script-writing functions: 
\code{\link{scriptAPIM}()},
\code{\link{scriptCFA}()},
\code{\link{scriptCFM}()},
\code{\link{scriptINULL}()},
\code{\link{scriptISAT}()},
\code{\link{scriptMIM}()}
}
\concept{script-writing functions}
