% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scriptCor.R
\name{scriptCor}
\alias{scriptCor}
\title{A Function That Writes, Saves, and Exports Syntax for
Fitting Correlated Dyadic Factor Models}
\usage{
scriptCor(
  dvn,
  scaleset = "FF",
  lvname = "X",
  lvar = "X",
  constr_dy_meas = c("loadings", "intercepts", "residuals"),
  constr_dy_struct = c("variances", "means"),
  writeTo = NULL,
  fileName = NULL,
  outputType = "lavaan script"
)
}
\arguments{
\item{dvn}{Input dvn list from \code{scrapeVarCross()}}

\item{scaleset}{Input character to specify how to set the scale of the latent variable. Default is
\code{"FF"} (fixed-factor; see Details for rationale), but user can specify \code{"MV"} (Marker Variable)}

\item{lvname}{Input character to (arbitrarily) name the latent variable in \code{lavaan} syntax}

\item{lvar}{Input character to specify whether the latent variable represents
"X" or "Y" in the model. Default is \code{"X"}. This argument controls parameter
labeling (e.g., \code{lx} vs \code{ly} for loadings, \code{tx} vs \code{ty} for intercepts).}

\item{constr_dy_meas}{Input character vector detailing which measurement model parameters to constrain across dyad members.}

\item{constr_dy_struct}{Input character vector detailing which structural model parameters to constrain across dyad members.
Default is \code{c("variances", "means")}(in combination with defaults for \code{constr_dy_meas}, an indistinguishable correlated dyadic factors model),
but user can specify any combination of \code{"variances"} and \code{"means"}, or \code{"none"}.}

\item{writeTo}{A character string specifying a directory path to where the
output file(s) should be saved. If set to \code{"."}, the file(s) will be written
to the current working directory. The default is \code{NULL}, and examples use a
temporary directory created by \code{tempdir()}. When dealing with tabular output,
\code{writeTo} is only relevant if \code{gtTab = TRUE}.}

\item{fileName}{A character string specifying a desired base name for the
output file. The default is \code{NULL}. The specified name will be automatically
appended with the appropriate file extension (e.g., \code{.txt} for \code{lavaan}
scripts, \code{.rtf} for tabular output when \code{gtTab = TRUE}, or other extensions
as appropriate for the output type). If a file with the same name already
exists in the user's chosen directory, it will be overwritten.}

\item{outputType}{Character string specifying the type of output to return.
Options are \code{"lavaan script"} (default) to return a character object of
\code{lavaan} syntax that can be passed immediately to \code{lavaan} functions, or
\code{"syntax components"} to return a structured list of model components.}
}
\value{
Character object of \code{lavaan} script that can be passed immediately to
\code{lavaan} functions (when \code{outputType = "lavaan script"}), or a structured list
of model components (when \code{outputType = "syntax components"}).
}
\description{
This function takes the outputted object from \code{scrapeVarCross()}
and automatically writes, returns, and exports (\code{.txt}) \code{lavaan} syntax
for specifying dyadic configural, loading, intercept, and residual invariant
two-factor models.
}
\details{
\itemize{
\item By default, many \verb{dySEM::} functions (including \code{scriptCor()}) default to
a fixed-factor method of scale-setting, whereby the latent variance of
a given factor is constrained to 1 for both partners in the configurally invariant
model, and then one of these variances is freely estimated in subsequent
models of the invariance testing sequence.
We have selected this default for two reasons:
(1) the selection of a marker-variable is usually arbitrary,
yet can have a large influence on the estimation and testing of of structural parameters
(see https://stats.stackexchange.com/questions/402133/in-cfa-does-it-matter-which-factor-loading-is-set-to-1/402732#402732);
and (2) the selection of a non-invariant marker-variable
can have disastrous down-stream consequences for the identification of
non-invariant measurement parameters, following a the rejection of an omnibus
invariance constraint set (see Lee, Preacher, & Little, 2011).
}
}
\examples{
dvn <- scrapeVarCross(
  dat = commitmentQ,
  x_order = "spi",
  x_stem = "sat.g",
  x_delim1 = ".",
  x_delim2 = "_",
  distinguish_1 = "1",
  distinguish_2 = "2"
)

sat.indist.script <- scriptCor(
  dvn,
  scaleset = "FF",
  lvname = "Sat"
)

sat.lvars.script <- scriptCor(
  dvn,
  scaleset = "FF",
  lvname = "Sat",
  constr_dy_meas = "loadings",
  constr_dy_struct = "variances"
)

sat.resids.script <- scriptCor(
  dvn,
  scaleset = "FF",
  lvname = "Sat",
  constr_dy_meas = c("loadings", "intercepts", "residuals"),
  constr_dy_struct = "none",
  writeTo = tempdir(),
  fileName = "dCor_residual"
)

sat.ints.script <- scriptCor(
  dvn,
  scaleset = "FF",
  lvname = "Sat",
  constr_dy_meas = c("loadings", "intercepts"),
  constr_dy_struct = "none",
  writeTo = tempdir(),
  fileName = "dCor_intercept"
)

sat.loads.script <- scriptCor(
  dvn,
  scaleset = "FF",
  lvname = "Sat",
  constr_dy_meas = c("loadings"),
  constr_dy_struct = "none",
  writeTo = tempdir(),
  fileName = "dCor_loading"
)

sat.config.script <- scriptCor(
  dvn,
  scaleset = "FF",
  lvname = "Sat",
  constr_dy_meas = "none",
  constr_dy_struct = "none",
  writeTo = tempdir(),
  fileName = "dCor_configural"
)
}
\seealso{
\code{\link{scrapeVarCross}} which this function relies on.

Other uni-construct script-writing functions: 
\code{\link{scriptBifac}()},
\code{\link{scriptHier}()},
\code{\link{scriptUni}()}
}
\concept{uni-construct script-writing functions}
