% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scriptHelpers.R
\name{scriptHelpers}
\alias{scriptHelpers}
\alias{loads}
\alias{intercepts}
\alias{resids}
\alias{lvars}
\alias{lmeans}
\title{Helper-functions for scripting free, fixed, and equated families of parameters}
\usage{
loads(dvn, lvar = "X", lvname, partner = "1", type = "free")

intercepts(dvn, lvar = "X", partner = "1", type = "free")

resids(dvn, lvar = "X", partner = "1", type = "free")

lvars(dvn, lvar = "X", lvname, partner = "1", type = "free")

lmeans(dvn, lvar = "X", lvname, partner = "1", type = "free")
}
\arguments{
\item{dvn}{input dvn list from scrapeVarCross}

\item{lvar}{input character for whether scripting helpers target latent "X" or :Y" indicator variables in dvn}

\item{lvname}{input character to (arbitrarily) name LV in lavaan syntax}

\item{partner}{input character to indicate parameters for first or second dyad member}

\item{type}{input character to indicate whether parameters "fixed", "free", "equated", or "equated_mv" in estimation}
}
\description{
Helper-functions for scripting free, fixed, and equated families of parameters
}
\seealso{
Other helpers: 
\code{\link{esHelpers}},
\code{\link{outHelpers}},
\code{\link{outParamHelpers}},
\code{\link{scrapeHelpers}},
\code{\link{semHelpers}}
}
\concept{helpers}
