% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_funs.R
\name{sem_funs}
\alias{sem_funs}
\alias{lregs}
\alias{cfloads}
\alias{cfvars}
\alias{cfmeans}
\title{Functions for scripting structural parameters}
\usage{
lregs(dvn, param, lvxname, lvyname, type = "free")

cfloads(dvn, lvxname, lvyname, type = "equated")

cfvars(lvname, type)

cfmeans(lvname, type)
}
\arguments{
\item{dvn}{input object from dyadVarNames()}

\item{param}{input character for what kind of parameter is being scripted ("act", "apim_part", "mim_part", "cf")}

\item{lvxname}{input character to (arbitrarily) name LV X in lavaan syntax}

\item{lvyname}{input character to (arbitrarily) name LV Y in lavaan syntax}

\item{type}{input character to indicate whether parameters "free" or "equated" in estimation}

\item{lvar}{input character for whether scripting helpers target latent "X" or :Y" indicator variables in dvn}

\item{partner}{input character to indicate parameters for first or second dyad member}
}
\description{
Functions for scripting structural parameters
}
\seealso{
\code{\link{dyadCFA}} which this function supplies
}
