% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siptExtractor.R
\name{siptExtractor}
\alias{siptExtractor}
\title{A helper function for scraping variable names from (S)tem-(I)tem-(P)artner-(T)ime
formatted variables}
\usage{
siptExtractor(dat, stem, delim1, item_num, delim2, distinguish, delim3, wave)
}
\arguments{
\item{dat}{input data frame of indicators of a particular LV}

\item{stem}{input character stem of indicator variables for LV}

\item{delim1}{optional input character of first delimiting character}

\item{item_num}{optional input character to indicate range of variable numbers (defaults to scrape any digits)}

\item{delim2}{optional input character of second delimiting character}

\item{distinguish}{input character used as the identifier a particular partner}

\item{delim3}{optional input character of second delimiting character}

\item{wave}{input character for the particular wave of a longitudinal design}
}
\value{
vector containing variable names for a particular partner, for
a particular LV,
}
\description{
This function helps dvn extract variable names according to the
"sip" order type.
}
