% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transDtoP.R
\name{transDtoP}
\alias{transDtoP}
\title{A function for transforming (D)yad datasets in(to) (P)airwise datasets (UPDATE EXAMPLE)}
\usage{
transDtoP(dat, firstvar1, lastvar1, firstvar2, lastvar2)
}
\arguments{
\item{dat}{input data frame of indicators of a particular LV}

\item{firstvar1}{input character for first variable/column to extract for P1}

\item{lastvar1}{input character for last variable/column to extract for P1}

\item{firstvar2}{input character for first variable/column to extract for P2}

\item{lastvar2}{input character for last variable/column to extract for P2}
}
\value{
Pairwise-structured data frame
}
\description{
This function transforms dyad-styled datasets into pairwise. The function
presumes that variables/columns for a given partner are blocked together in
a dataframe (e.g., P1 variables followed by P2)
}
\examples{
dvn <- dvn2(dat = dat.pkg, x_order = "sip", x_stem = "X", x_delim2=".", x_item_num="\\\\d", distinguish_1="1", distinguish_2="2",
y_order="sip", y_stem="Y", y_delim2=".", y_item_num="\\\\d")
}
