% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transPtoD.R
\name{transPtoD}
\alias{transPtoD}
\title{A function for transforming (P)airwise datasets in(to) (D)yad datasets (UPDATE EXAMPLE)}
\usage{
transPtoD(dat)
}
\arguments{
\item{dat}{input data frame of indicators of a particular LV}
}
\value{
Dyad-structured data frame
}
\description{
This function transforms pairwise-styled datasets into dyad T
}
\examples{
dvn <- dvn2(dat = dat.pkg, x_order = "sip", x_stem = "X", x_delim2=".", x_item_num="\\\\d", distinguish_1="1", distinguish_2="2",
y_order="sip", y_stem="Y", y_delim2=".", y_item_num="\\\\d")
}
