% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaanscript_funs.R, R/varextract_funs.R
\name{x2vars}
\alias{x2vars}
\alias{y1vars}
\alias{y2vars}
\alias{varextract_funs}
\alias{x1vars}
\title{Functions for dyadVarNames() for extracting item information}
\usage{
x2vars(dat, index2, xvar)

y1vars(dat, index1, yvar)

y2vars(dat, index2, yvar)

x1vars(dat, index1, xvar)

x2vars(dat, index2, xvar)

y1vars(dat, index1, yvar)

y2vars(dat, index2, yvar)
}
\arguments{
\item{dat}{input data frame of indicators of a particular LV}

\item{index2}{combined character of sep and identifier for the first partner}

\item{xvar}{input character stem of indicator variables for LV X}

\item{index1}{combined character of sep and identifier for the first partner}

\item{yvar}{input character stem of indicator variables for LV Y}
}
\description{
Functions for dyadVarNames() for extracting item information
}
\seealso{
\code{\link{dyadVarNames}} which this function supplies
}
