% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfa_funs.R
\name{cfa_funs}
\alias{cfa_funs}
\alias{loads}
\alias{intercepts}
\alias{resids}
\alias{lvars}
\alias{lmeans}
\title{Functions for dyadCFA() for scripting free or fixed families of parameters}
\usage{
loads(dvn, lvname, partner = "1", type = "free")

intercepts(dvn, partner = "1", type = "free")

resids(dvn, partner = "1", type = "free")

lvars(dvn, lvname, partner = "1", type = "free")

lmeans(dvn, lvname, partner = "1", type = "free")
}
\arguments{
\item{dvn}{input object from dyadVarNames()}

\item{lvname}{input character to (arbitrarily) name LV in lavaan syntax}

\item{partner}{input character to indicate parameters for first or second dyad member}

\item{type}{input character to indicate whether parameters "fixed", "free" or "equated" in estimation}
}
\description{
Functions for dyadCFA() for scripting free or fixed families of parameters
}
\seealso{
\code{\link{dyadCFA}} which this function supplies
}
