% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRES-data.R
\docType{data}
\name{DRES}
\alias{DRES}
\title{Relationship quality and sexual satisfaction of 121 couples}
\format{
A data frame with 121 rows and 28 variables:
\describe{
\item{PRQC_1.1}{PRQC items for partner 1}
\item{PRQC_1.2}{PRQC items for partner 2}
\item{sexsat1.1}{sexual satisfaction items for partner 1}
\item{sexsat1.1}{sexual satisfaction items for partner 2}
}
}
\usage{
data(DRES)
}
\description{
A dataset containing 9 observed indicators of relationship quality (PRQC)
and 5 observed indicators of sexual satisfaction from 121 couples
in the DRES (Daily Relationship Experiences Study; Raposo, Impett, & Muise, in press)
}
\references{
Raposo, S., Impett, E. A., & Muise, A. (in Press). Avoidantly Attached Individuals Are More Exchange‐Oriented and Less Communal in the Bedroom. \emph{Archives of Sexual Behavior}. https://doi.org/10.1007/s10508-020-01813-9
}
\keyword{datasets}
