% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyadCFA.R, R/dyadCFA_exp.R
\name{dyadCFA}
\alias{dyadCFA}
\title{A Function That Writes, Saves, and Exports Syntax for
Fitting Latent Dyadic Measurement and Invariance Models}
\usage{
dyadCFA(dvn, lvname, model = "configural")

dyadCFA(dvn, lvname, model = "configural")
}
\arguments{
\item{dvn}{input object from dyadVarNames()}

\item{lvname}{input character to (arbitrarily) name LV in lavaan syntax}

\item{model}{input character used to specify which level of invariance is
modeled ("configural", "loading", "intercept", "residual", or "indistinguishable"). Defaults to "configural".}

\item{lvnum}{optional input character to indicate which LV is modeled ("one" or "two").
Only necessary if dvn contains both X and Y information and user wants CFA for Y}
}
\value{
character object of lavaan script that can be passed immediately to
lavaan functions

character object of lavaan script that can be passed immediately to
lavaan functions
}
\description{
This function takes the outputted object from dyadVarNames()
and automatically writes, returns, and exports (.txt) lavaan() syntax
for specifying dyadic configural, loading, and intercept invariant
measurement models for either a specified X or Y factor.

This function takes the outputted object from dyadVarNames()
and automatically writes, returns, and exports (.txt) lavaan() syntax
for specifying dyadic configural, loading, and intercept invariant
measurement models for either a specified X or Y factor.
}
\examples{
dvn = dyadVarNames(dat, xvar="X", sep = ".",distinguish1 = "1", distinguish2 = "2")
con.config.script = dyadCFA(dvn, lvname = "Conflict", model = "configural")
con.loading.script = dyadCFA(dvn, lvname = "Conflict",  model = "loading")
con.intercept.script = dyadCFA(dvn, lvname = "Conflict",  model = "intercept")
dvn = dyadVarNames(dat, xvar="X", sep = ".",distinguish1 = "1", distinguish2 = "2")
con.config.script = dyadCFA(dvn, lvname = "Conflict", model = "configural")
con.loading.script = dyadCFA(dvn, lvname = "Conflict",  model = "loading")
con.intercept.script = dyadCFA(dvn, lvname = "Conflict",  model = "intercept")
}
\seealso{
\code{\link{dyadVarNames}} which this function relies on

\code{\link{dyadVarNames}} which this function relies on
}
