% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esHelpers.R
\name{esHelpers}
\alias{esHelpers}
\alias{grouploads}
\alias{groupints}
\alias{groupindsds}
\alias{grouplvmean}
\alias{grouplvsd}
\title{Helper-functions for noninvariance effect size functions}
\usage{
grouploads(fit, dvn, source)

groupints(fit, dvn, source)

groupindsds(dat, dvn, source)

grouplvmean(fit, source)

grouplvsd(fit, source)
}
\arguments{
\item{fit}{outputted dyadic cfa lavaan object}

\item{dvn}{input dvn list from scrapeVarCross}

\item{source}{character for whether parameter of interest should be extracted for group "1", "2", or from the "nogroup" model}

\item{dat}{data frame of indicators}
}
\value{
a numeric vector containing the estimated values of the parameter of interest (e.g., loadings, intercepts, etc.) for the desired group.
}
\description{
Helper-functions for noninvariance effect size functions
}
\examples{
dvn <- scrapeVarCross(dat = commitmentQ, x_order = "spi", x_stem = "sat.g", x_delim1 = ".", 
x_delim2="_", distinguish_1="1", distinguish_2="2")
sat.config.script <-  scriptCFA(dvn, lvname = "Sat", model = "configural")
sat.config.mod <- lavaan::cfa(sat.config.script, data = commitmentQ, std.lv = FALSE, 
auto.fix.first= FALSE, meanstructure = TRUE)
grouploads(sat.config.mod, dvn, "2")
}
\concept{helpers}
