% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_funs.R
\name{out_funs}
\alias{out_funs}
\alias{makeTable}
\title{Functions for scripting structural parameters}
\usage{
makeTable(dvn, fit, model, tabletype)
}
\arguments{
\item{dvn}{input object from dyadVarNames()}

\item{param}{input character for what kind of parameter is being scripted ("act", "apim_part", "mim_part", "cf")}

\item{lvar}{input character for whether scripting helpers target latent "X" or :Y" indicator variables in dvn}

\item{lvxname}{input character to (arbitrarily) name LV X in lavaan syntax}

\item{lvyname}{input character to (arbitrarily) name LV Y in lavaan syntax}

\item{partner}{input character to indicate parameters for first or second dyad member}

\item{type}{input character to indicate whether parameters "free" or "equated" in estimation}
}
\description{
Functions for scripting structural parameters
}
\seealso{
\code{\link{dyadCFA}} which this function supplies
}
