% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scriptAPIM.R
\name{scriptAPIM}
\alias{scriptAPIM}
\title{A Function That Writes, Saves, and Exports Syntax for
Fitting Latent Actor-Partner Interdependence Models}
\usage{
scriptAPIM(
  dvn,
  lvxname,
  lvyname,
  model = "configural",
  equate = "none",
  k = FALSE
)
}
\arguments{
\item{dvn}{input object from dyadVarNames()}

\item{lvxname}{input character to (arbitrarily) name LV X in lavaan syntax}

\item{lvyname}{input character to (arbitrarily) name LV Y in lavaan syntax}

\item{model}{input character used to specify which level of invariance is
modeled. Defaults to "configural"}

\item{equate}{input character to specify which type of structural parameters
("actor" = actor effects, "partner" = partner effects,
"all_effects" = actor and partner effects, x_means, y_means, or all_means)
are constrained to equivalencey between partners. Default is "none". Actor or partner effect constraints
require at least a loading-invariant model to be specified, otherwise a warning is returned; mean constraints
require an intercept-invariant model to be specified, otherwise a warning is returned.}

\item{k}{input logical for whether Kenny & Ledermann's (2010) k parameter should be
calculated to characterize the dyadic pattern in the APIM. Defaults FALSE, and requires at least
a loading-invariant model to be specified, otherwise a warning is returned.}
}
\value{
character object of lavaan script that can be passed immediately to
lavaan functions. Users will receive message if structural comparisons are specified
when the recommended level of invariance is not also specified. If user supplies dvn
with containing X or Y variables, they are alerted to respecify the dvn object.
}
\description{
This function takes the outputted object from dyadVarNames()
and automatically writes, returns, and exports (.txt) lavaan() syntax
for specifying Actor-Partner Interdependence Models (APIMs). Users can
also invoke configural, loading, and/or intercept invariant
measurement models, and particular types of structural comparisons.
}
\examples{
dvn = dyadVarNames(dat, xvar="X", yvar="Y", sep = ".",distinguish1 = "1", distinguish2 = "2")
apim.script.config = scriptAPIM(dvn, lvxname = "Conflict",
lvyname = "Satisfaction", model = "configural")
apim.script.load = scriptAPIM(dvn, lvxname = "Conflict",
lvyname = "Satisfaction", model = "loading")
apim.script.int = scriptAPIM(dvn, lvxname = "Conflict",
lvyname = "Satisfaction", model = "intercept")
apim.script.load.actor = scriptAPIM(dvn, lvxname = "Conflict",
lvyname = "Satisfaction", model = "loading", equate = "actor")
}
\seealso{
\code{\link{dyadVarNames}} which this function relies on

Other script-writing functions: 
\code{\link{scriptBifactor}()},
\code{\link{scriptCFA}()},
\code{\link{scriptCFM}()},
\code{\link{scriptINULL}()},
\code{\link{scriptISAT}()}
}
\concept{script-writing functions}
