% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scriptCFM.R
\name{scriptCFM}
\alias{scriptCFM}
\title{A Function That Writes, Saves, and Exports Syntax for
Fitting Latent Common Fate Models}
\usage{
scriptCFM(
  dvn,
  lvxname,
  lvyname,
  model = "indist",
  scaleset = "FF",
  writescript = FALSE
)
}
\arguments{
\item{dvn}{input dvn list from scrapeVarCross}

\item{lvxname}{input character to (arbitrarily) name LV X in lavaan syntax}

\item{lvyname}{input character to (arbitrarily) name LV Y in lavaan syntax}

\item{model}{input character used to specify which level of invariance is
modeled. Defaults to "configural"}

\item{scaleset}{input character to specify how to set the scale of the latent variable(s). Default is
"FF" (fixed-factor; see Details for rationale), but user can specify "MV" (Marker Variable)}

\item{writescript}{input logical (default FALSE) for whether lavaan script should
be concatenated and written to current working directory (in subdirectory "scripts")}
}
\value{
character object of lavaan script that can be passed immediately to
lavaan functions. Users will receive message if structural comparisons are specified
when the recommended level of invariance is not also specified. If user supplies dvn
with containing X or Y variables, they are alerted to respecify the dvn object.
}
\description{
This function takes the outputted object from scrapeVarCross()
and automatically writes, returns, and exports (.txt) lavaan() syntax
for specifying Common Fate Models (CFMs). Users can
also invoke configural, loading, and/or intercept invariant
measurement models, and particular types of structural comparisons.
}
\examples{
dvn <- scrapeVarCross(dat = DRES, x_order = "sip", x_stem = "PRQC", x_delim1 = "_",
x_delim2=".", distinguish_1="1", distinguish_2="2",
y_order="sip", y_stem="sexsat", y_delim2=".")
cfm.script.config  <-  scriptCFM(dvn, lvxname = "Quality", lvyname = "SexSat",
model = "configural", scaleset = "FF")
cfm.script.load  <-  scriptCFM(dvn, lvxname = "Quality", lvyname = "SexSat",
model = "loading", scaleset = "FF")
cfm.script.int  <-  scriptCFM(dvn, lvxname = "Quality", lvyname = "SexSat",
model = "intercept", scaleset = "FF")
cfm.script.ind  <-  scriptCFM(dvn, lvxname = "Quality", lvyname = "SexSat",
model = "indist", scaleset = "FF")
}
\seealso{
\code{\link{scrapeVarCross}} which this function relies on

Other script-writing functions: 
\code{\link{scriptAPIM}()},
\code{\link{scriptBiDy}()},
\code{\link{scriptCFA}()},
\code{\link{scriptINULL}()},
\code{\link{scriptISAT}()},
\code{\link{scriptMIM}()}
}
\concept{script-writing functions}
