% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scriptISAT.R
\name{scriptISAT}
\alias{scriptISAT}
\title{A Function That Writes, Saves, and Exports Syntax for
Fitting the I-SAT model for indistinguishable dyads}
\usage{
scriptISAT(dvn, lvxname = "X", lvyname = NULL, writescript = FALSE)
}
\arguments{
\item{dvn}{input dvn list from scrapeVarCross}

\item{lvxname}{input character to (arbitrarily) name X LV in lavaan syntax}

\item{lvyname}{(optional) input character to (arbitrarily) name X LV in lavaan syntax}

\item{writescript}{input logical (default FALSE) for whether lavaan script should
be concatenated and written to current working directory}
}
\value{
character object of lavaan script that can be passed immediately to
lavaan functions
}
\description{
This function takes the outputted object from scrapeVarCross()
and automatically writes, returns, and exports (.txt) lavaan() syntax
for the I-SAT model described in Olsen & Kenny (2006)
}
\examples{
dvn <- scrapeVarCross(dat = DRES, x_order = "sip", x_stem = "PRQC", x_delim1 = "_",
x_delim2=".", x_item_num="\\\\d+", distinguish_1="1", distinguish_2="2")
qual.isat.script <- scriptISAT(dvn, lvxname = "Qual")
}
\seealso{
\code{\link{scrapeVarCross}} which this function relies on

Other script-writing functions: 
\code{\link{scriptAPIM}()},
\code{\link{scriptBiDy}()},
\code{\link{scriptCFA}()},
\code{\link{scriptCFM}()},
\code{\link{scriptINULL}()},
\code{\link{scriptMIM}()}
}
\concept{script-writing functions}
