% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scriptISAT.R
\name{scriptISAT}
\alias{scriptISAT}
\title{A Function That Writes, Saves, and Exports Syntax for
Fitting the I-SAT model for indistinguishable dyads}
\usage{
scriptISAT(dvn, lvname = "X")
}
\arguments{
\item{dvn}{input object from dyadVarNames()}

\item{lvname}{input character to (arbitrarily) name LV in lavaan syntax}
}
\value{
character object of lavaan script that can be passed immediately to
lavaan functions
}
\description{
This function takes the outputted object from dyadVarNames()
and automatically writes, returns, and exports (.txt) lavaan() syntax
for the I-SAT model described in Olsen & Kenny (2006)
}
\examples{
dvn = dyadVarNames(dat, xvar="X", sep = ".",distinguish1 = "1", distinguish2 = "2")
isat.mod = scriptISAT(dvn, lvname = "X")
}
\seealso{
\code{\link{dyadVarNames}} which this function relies on

Other script-writing functions: 
\code{\link{scriptAPIM}()},
\code{\link{scriptBifactor}()},
\code{\link{scriptCFA}()},
\code{\link{scriptCFM}()},
\code{\link{scriptINULL}()}
}
\concept{script-writing functions}
