% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scriptObsAPIM.R
\name{scriptObsAPIM}
\alias{scriptObsAPIM}
\title{A Function That Writes, Saves, and Exports Syntax for
Fitting Observed Actor-Partner Interdependence Models}
\usage{
scriptObsAPIM(
  X1 = NULL,
  Y1 = NULL,
  X2 = NULL,
  Y2 = NULL,
  equate = "none",
  k = FALSE,
  writescript = FALSE
)
}
\arguments{
\item{X1}{character of vector name containing X variable/composite for partner 1}

\item{Y1}{character of vector name containing Y variable/composite for partner 1}

\item{X2}{character of vector name containing X variable/composite for partner 2}

\item{Y2}{character of vector name containing Y variable/composite for partner 2}

\item{equate}{character of what parameter(s) to constrain ("actor", "partner", "all"); default is "none" (all freely estimated)}

\item{k}{input logical for whether Kenny & Ledermann's (2010) k parameter should be
calculated to characterize the dyadic pattern in the APIM. Default to FALSE}

\item{writescript}{input logical (default FALSE) for whether lavaan script should
be concatenated and written to current working directory}
}
\value{
character object of lavaan script that can be passed immediately to
lavaan functions.
}
\description{
A Function That Writes, Saves, and Exports Syntax for
Fitting Observed Actor-Partner Interdependence Models
}
\examples{

obsAPIMScript <- scriptObsAPIM (X1 = "SexSatA", Y1 = "RelSatA",
X2 = "SexSatB", Y2 = "RelSatB",
equate = "none")

}
